menu refactor
